- move types to types directory - update some types to use LevelRange instead of number
