+++ src/Equipment/Weapon/weapons/maces.ts
@@ -1,6 +1,6 @@
-import AttackType from '../../../AttackType';
+import AttackType from '../../../types/AttackType';
 import StatType from '../../../Character/Stats/StatType';
-import DamageType from '../../../DamageType';
+import DamageType from '../../../types/DamageType';
 import { ItemType } from '../../Item';
 import { type Weapon, WeaponType } from '../Weapon';
