+++ src/Equipment/Weapon/weapons/wands.ts
@@ -1,7 +1,7 @@
-import AttackType from '../../../AttackType';
+import AttackType from '../../../types/AttackType';
 import AttributeType from '../../../Character/Attributes/AttributeType';
 import StatType from '../../../Character/Stats/StatType';
-import DamageType from '../../../DamageType';
+import DamageType from '../../../types/DamageType';
 import { ItemType } from '../../Item';
 import { type Weapon, WeaponType } from '../Weapon';
