+++ src/types/HitType.ts
