+++ src/types/LevelRange.ts
@@ -1,0 +1,3 @@
+type LevelRange = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13 | 14 | 15 | 16 | 17 | 18 | 19 | 20;
+
+export default LevelRange;
